Created SleepHelper class, moving safe_sleep, short_hybernate, long_hybernate, and listen_loiter into the SleepHelper class
